Revert "ONPREM-43: Uninstalling with preserve data option will change ownership of data directories to install user from hadoop user"
This reverts commit eccd3e114287468d6526d5df1bab590c3ea003dd.

+++ src/packages/win/multi-scripts/uninstall.ps1
@@ -43,20 +43,10 @@ function Main
     foreach ($folder in ${ENV:HDFS_DATA_DIR}.Split(","))
     {
         $folder = $folder.Trim()
-        if ($folder -ne $null)
+        if ( ($folder -ne $null) -and (($forceclean -eq $true) -or (-not (Test-Path "$folder\*"))) )
         {
-            if ( ($forceclean -eq $true) -or (-not (Test-Path "$folder\*")) )
-            {
-                $cmd = "rd /s /q `"$folder`""
-                Invoke-Cmd $cmd
-            }
-            else
-            {
-                if (Test-Path "$folder\*")
-                {
-                    GiveFullPermissions $folder $ENV:USERNAME $true
-                }
-            }
+            $cmd = "rd /s /q `"$folder`""
+            Invoke-Cmd $cmd
         }
     }
